feat : 교환 패키지 세분화
