+++ src/main/java/com/barter/domain/trade/donationtrade/entity/DonationTrade.java
@@ -1,4 +1,4 @@
-package com.barter.domain.trade.entity;
+package com.barter.domain.trade.donationtrade.entity;
 
 import com.barter.domain.BaseTimeStampEntity;
 import com.barter.domain.product.entity.RegisteredProduct;
