+++ src/main/java/com/barter/domain/trade/donationtrade/service/DonationTradeService.java
@@ -1,4 +1,4 @@
-package com.barter.domain.trade.service;
+package com.barter.domain.trade.donationtrade.service;
 
 import org.springframework.stereotype.Service;
