+++ src/main/java/com/barter/domain/trade/periodtrade/controller/PeriodTradeController.java
@@ -1,4 +1,4 @@
-package com.barter.domain.trade.controller;
+package com.barter.domain.trade.periodtrade.controller;
 
 import org.springframework.web.bind.annotation.RestController;
 
@@ -6,5 +6,5 @@ import lombok.RequiredArgsConstructor;
 
 @RestController
 @RequiredArgsConstructor
-public class DonationTradeController {
+public class PeriodTradeController {
 }
